Moved fucntions to the user class not tested yet
